+++ frontend/components/DeleteItem.js
@@ -31,6 +31,9 @@ class DeleteItem extends Component {
        {(deleteItem, { error }) => (
           <button
             onClick={() => {
+              if(!this.props.me){
+                return alert("Please sign in to delete an item.")
+              }
               if (confirm('Are you sure you want to delete this item?')) {
                 deleteItem().catch(err => {
                   alert(err.message);
